+++ CHANGELOG.md
@@ -10,6 +10,7 @@ All notable changes to this project will be documented in this file.
 - Bugged file_to_zone round trip, added test
 - Removed unnecessary np.unique() call in zone construction
 - Fixed function definition in Lamp.get_polar() and Lamp.get_cartesian()
+- Lamp equality handled even if they have no photometry
 
 ### Changed